implemented fuel and resolved warnings during build

+++ src/ECS/TextureManager.cpp
@@ -16,7 +16,7 @@ void TextureManager::Draw(SDL_Texture* tex, SDL_Rect src, SDL_Rect dest, SDL_Ren
 void TextureManager::DrawLine(float startX, float startY, float endX, float endY) {
 
 	SDL_SetRenderDrawColor(Game::renderer, 255, 255, 255, SDL_ALPHA_OPAQUE);
-	SDL_RenderDrawLine(Game::renderer, startX, startY, endX, endY);
+	SDL_RenderDrawLine(Game::renderer, (int) startX, (int) startY, (int) endX, (int) endY);
 	SDL_SetRenderDrawColor(Game::renderer, 19, 24, 98, 255);
 }
